+++ actix-web/src/http/mod.rs
@@ -3,4 +3,4 @@
 pub mod header;
 
 // TODO: figure out how best to expose http::Error vs actix_http::Error
-pub use actix_http::{uri, ConnectionType, Error, Method, StatusCode, Uri, Version};
+pub use actix_http::{uri, ConnectionType, Error, KeepAlive, Method, StatusCode, Uri, Version};