+++ actix-web/CHANGES.md
@@ -7,6 +7,7 @@
 
 [#2625]: https://github.com/actix/actix-web/pull/2625
 
+
 ## 4.0.0-rc.2 - 2022-02-02
 ### Added
 - On-by-default `macros` feature flag to enable routing and runtime macros. [#2619]