+++ actix-web/CHANGES.md
@@ -1,6 +1,11 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+### Added
+- Implement `Responder` for `Vec<u8>`. [#2625]
+- Re-export `KeepAlive` in `http` mod. [#2625]
+
+[#2625]: https://github.com/actix/actix-web/pull/2625
 
 
 ## 4.0.0-rc.2 - 2022-02-02