+++ actix-web/CHANGES.md
@@ -4,9 +4,10 @@
 ### Added
 - Implement `Responder` for `Vec<u8>`. [#2625]
 - Re-export `KeepAlive` in `http` mod. [#2625]
+- `middleware::logger::custom_status_replace`. [#2631]
 
 [#2625]: https://github.com/actix/actix-web/pull/2625
-
+[#2631]: https://github.com/actix/actix-web/pull/2631
 
 ## 4.0.0-rc.2 - 2022-02-02
 ### Added