+++ actix-web/CHANGES.md
@@ -1,6 +1,9 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+
+
+## 4.0.0-rc.3 - 2022-02-08
 ### Changed
 - `middleware::Condition` gained a broader compatibility; `Compat` is needed in fewer cases. [#2635]